[test] Improve cleanup of tests that use schedulers / processors
In order to avoid side-effects and dangling threads, this big commit
adds cleanup steps to a lot of tests that instantiate a new Scheduler,
ExecutorService or even Processor (as eg. the EventLoopProcessors will
spawn threads).

As often as possible, a thread factory is used so that the thread names
will reflect the test from which they were created.

Additional changes for debugging of #1028 include thread factories that
dump a stacktrace in the created thread's name and a thread dump on
failure of the 2 FluxGroupBy tests. And said tests use 2 separate
`Scheduler`.